+++ testing/lutinfobox.js
@@ -729,6 +729,9 @@ LUTInfoBox.prototype.createClgInfo = function() {
 	this.insClg.appendChild(this.insClgBack);
 	this.insClgInfo = document.createElement('div');
 	this.insClgInfo.setAttribute('class','infotext');
+	this.addInfo(this.insClgInfo,false,null,"v3.3 Beta 1");
+	this.addInfo(this.insClgInfo,true,null,"* New Feature - Additional cameras; Sony, Panasonic, DJI.");
+	this.addInfo(this.insClgInfo,true,null,"* New Feature - Sony Venice-tuned S-Gamut3 and S-Gamut3.cine using primaries calculated from Sony's IDTs (these shouldn't be camera-specific, the S-Gamut primaries should be set in stone, but Sony definitely offers different ones for the Venice).");
 	this.addInfo(this.insClgInfo,false,null,"v3.2 Beta 1");
 	this.addInfo(this.insClgInfo,true,null,"New Feature - DJI DLog and DGamut (Zenmuse X5 and X7 version).");
 	this.addInfo(this.insClgInfo,true,null,"New Feature - DJI DLog-M based upon test images taken by Mark Walter a Mavic 2.");