+++ testing/splash.js
@@ -195,7 +195,7 @@ function updateSplash() {
 
 /********************** Start things up **********************/
 var lutInputs = new LUTInputs();
-lutInputs.addInput('version','v3.2 Beta 1');
+lutInputs.addInput('version','v3.3 Beta 1');
 lutInputs.addInput('date','April 2019');
 var splash = splashStart();
 var lutTests = new LUTTests(lutInputs);