+++ testing/CHANGELOG.md
@@ -1,6 +1,14 @@
 LUTCalc Changelog
 =================
 
+v3.3 Beta 1
+-------------
+* New Feature - Additional cameras; Sony, Panasonic, DJI, Nikon.
+* New Feature - Sony Venice-tuned S-Gamut3 and S-Gamut3.cine using primaries calculated from Sony's IDTs (these shouldn't be camera-specific, the S-Gamut primaries should be set in stone, but Sony definitely offers different ones for the Venice).
+* New Feature - Added Nikon N-Log gamma curve.
+* New Feature - Fixed point precision setting within settings, so that LUT files can be more precise than the default six or eight decimal places. Good for scene linear output LUTs.
+* Bugfix - Fixed Lumetri / Speedgrade cube linear and conventional gamma settings.
+
 v3.2 Beta 1
 -------------
 * New Feature - DJI DLog and DGamut (Zenmuse X5 and X7 version).